feat: improve percentage convertible
Rename PercentageRepresentable to PercentageConvertible
